+++ packages/core/src/errors/formattingErrors.ts
@@ -1,0 +1,2 @@
+export const createInvalidCutoffStyleError = (style: string) =>
+  `generaltranslation Formatting Error: Invalid cutoff style: ${style}.`;