+++ packages/next/src/server-dir/buildtime/getTranslations.tsx
@@ -127,10 +127,15 @@ export async function getTranslations(id?: string): Promise<
     ) => {
       try {
         // (1) Try to format message
-        return gt.formatMessage(message, {
+        const formattedMessage = gt.formatMessage(message, {
           locales,
           variables: options,
         });
+        const cutoffMessage = gt.formatCutoff(formattedMessage, {
+          locales,
+          maxChars: metadata?.$maxChars ?? options.$maxChars,
+        });
+        return cutoffMessage;
       } catch (error) {
         if (process.env.NODE_ENV === 'production') {
           console.warn(
@@ -158,7 +163,11 @@ export async function getTranslations(id?: string): Promise<
         }
 
         // (3) Fallback to original message (unformatted)
-        return message; // fallback to original message (unformatted)}
+        const cutoffMessage = gt.formatCutoff(message, {
+          locales,
+          maxChars: metadata?.$maxChars ?? options.$maxChars,
+        });
+        return cutoffMessage; // fallback to original message (unformatted)}
       }
     };
 
@@ -200,7 +209,9 @@ export async function getTranslations(id?: string): Promise<
       const hash = hashSource({
         source: entry,
         ...(metadata?.$context && { context: metadata.$context }),
-        ...(metadata?.$maxChars && { maxChars: metadata.$maxChars }),
+        ...(metadata?.$maxChars != null && {
+          maxChars: Math.abs(metadata.$maxChars),
+        }),
         id,
         dataFormat: 'ICU',
       });