+++ packages/react-core/src/provider/hooks/translation/useCreateInternalUseTranslationsObjFunction.ts
@@ -135,7 +135,7 @@ export function useCreateInternalUseTranslationsObjFunction(
                   targetLocale: locale,
                   metadata: {
                     ...(metadata?.$context && { context: metadata.$context }),
-                    ...(metadata?.$maxChars && {
+                    ...(metadata?.$maxChars != null && {
                       maxChars: metadata.$maxChars,
                     }),
                     id,