+++ packages/react-core/src/provider/hooks/useRuntimeTranslation.ts
@@ -339,14 +339,24 @@ export default function useRuntimeTranslation({
               ? {
                   dataFormat: 'JSX',
                   source: params.source as JsxChildren,
-                  metadata: params.metadata,
+                  metadata: {
+                    ...params.metadata,
+                    ...(params.metadata.maxChars != null && {
+                      maxChars: Math.abs(params.metadata.maxChars),
+                    }),
+                  },
                   resolve,
                   reject: () => {}, // no-op; we no longer reject
                 }
               : {
                   dataFormat: 'ICU',
                   source: params.source as string,
-                  metadata: params.metadata,
+                  metadata: {
+                    ...params.metadata,
+                    ...(params.metadata.maxChars != null && {
+                      maxChars: Math.abs(params.metadata.maxChars),
+                    }),
+                  },
                   resolve,
                   reject: () => {}, // no-op; we no longer reject
                 };